Fix image and quote refrence alignment
The references are now approximately aligned with the right side of a
quote or image. Right now this required a couple of hacks and magic
numbers and therefore doesn't work in all cases.

This commit also adds a shadow CSS class for images to make the size of
an image more visible if it has no background color.